+++ protobuf/CHANGELOG.md
@@ -4,6 +4,9 @@
   added to convert the message into the [official protocol buffers text
   format][text format]. ([#1080], [#125])
 
+* Add [well-known proto types][wkts] as libraries. This change is required for
+  protoc_plugin-25.0.0. ([#1081])
+
 * **Breaking:** Hide `PbList` and `PbMap` constructors. It is not possible to
   construct these values correctly in user code, so the constructors are now
   private. Existing uses of `PbList` can be replaced by `List` and `PbMap` can
@@ -14,6 +17,8 @@
 [text format]: https://protobuf.dev/reference/protobuf/textformat-spec/
 [#1080]: https://github.com/google/protobuf.dart/pull/1080
 [#125]: https://github.com/google/protobuf.dart/issues/125
+[wkts]: https://protobuf.dev/reference/protobuf/google.protobuf
+[#1081]: https://github.com/google/protobuf.dart/pull/1081
 [#1072]: https://github.com/google/protobuf.dart/pull/1072
 
 ## 5.1.0